First pass at using the Jeesh, via Ender
Hooked `ender` calls into the build process for `gulp dev` and `gulp prod`
Included the `ender` module as a dev dep, though I think it will need to be
installed `-g` to work in this state. Gonna look into `ender`'s api so maybe we
won't need to tell people to install it. They'll have to install `gulp`,
though, so maybe no big.